Switch from Double to BigDecimal to prevent rounding errors

+++ src/main/scala/Checkout.scala
@@ -1,4 +1,4 @@
 
 object Checkout {
-  def checkout(items: List[Fruit]): Double = items.map(_.price).sum
+  def checkout(items: List[Fruit]): BigDecimal = items.map(_.price).sum
 }
